feat(P15-F04): extract IncomeForm into its own component
Moves the inline IncomeForm function, its option list, and field
mapping out of MonthlyPage.tsx into src/components/IncomeForm.tsx,
mirroring F03's ExpenseForm extraction. No behavior change.
